@@ -8,35 +8,32 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
+    "@vue/compat": "^3.2.47",
+    "@vue/compiler-sfc": "^3.2.47",
     "axios": "^1.3.4",
     "core-js": "^3.6.4",
     "currency-symbol-map": "^5.0.1",
     "firebase": "^9.6.7",
     "luxon": "^2.3.0",
+    "pinia": "^2.0.34",
     "register-service-worker": "^1.7.1",
     "socket.io-client": "^2.3.0",
-    "vue": "^2.6.11",
+    "vue": "^3.2.47",
     "vue-moment": "^4.1.0",
-    "vue-router": "^3.1.6",
+    "vue-router": "^4.1.6",
     "vue-socket.io": "^3.0.7",
     "vuefire": "^2.2.5",
-    "vuex": "^3.1.3",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^5.0.8",
     "@vue/cli-plugin-pwa": "^5.0.8",
     "@vue/cli-plugin-router": "^5.0.8",
-    "@vue/cli-plugin-unit-jest": "~5.0.0",
     "@vue/cli-plugin-vuex": "^5.0.8",
     "@vue/cli-service": "^5.0.8",
-    "@vue/test-utils": "^1.1.3",
-    "@vue/vue2-jest": "^27.0.0-alpha.2",
+    "@vue/test-utils": "^2.3.2",
     "babel-jest": "^27.0.6",
-    "chai": "^4.1.2",
-    "jest": "^27.0.5",
     "sass": "^1.58.3",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.11"
+    "sass-loader": "^8.0.2"
   }
 }
