@@ -8,30 +8,18 @@
       <img :src="require('@/assets/icons/ellipsis.svg')" />
     </div>
     <div class="description">{{ expense.description }}</div>
-    <div class="amount">
-      {{ getSymbolFromCurrency(expense.currency)
-      }}{{ expense.amount.toFixed(2) }}
-    </div>
-    <div class="exclude" v-if="expense.exclude && expense.exclude.length">
-      Exclude: {{ expense.exclude.join(', ') }}
-    </div>
+    <div class="amount">{{ getSymbolFromCurrency(expense.currency) }}{{ expense.amount.toFixed(2) }}</div>
+    <div class="exclude" v-if="expense.exclude && expense.exclude.length">Exclude: {{ expense.exclude.join(', ') }}</div>
   </div>
 </template>
 
-<script>
+<script setup>
 import { formatDate } from '@/modules/common/services/util.service';
 import getSymbolFromCurrency from 'currency-symbol-map';
-export default {
-  name: 'expense-preview',
-  props: {
-    expense: { type: Object, required: true },
-  },
-  methods: {
-    formatDate,
-    getSymbolFromCurrency,
-  },
-};
+
+const props = defineProps({
+  expense: { type: Object, required: true },
+});
 </script>
 
-<style>
-</style>
+<style></style>
