@@ -7,43 +7,25 @@
       <div class="page-header">
         <div class="title">{{ isEdit ? 'Edit' : 'Add' }} Member</div>
       </div>
-      <input
-        type="text"
-        class="form-input"
-        placeholder="Enter Name"
-        v-model="memberToEdit.name"
-      />
-      <input
-        type="text"
-        class="form-input"
-        placeholder="Enter Email"
-        v-model="memberToEdit.email"
-      />
+      <input type="text" class="form-input" placeholder="Enter Name" v-model="memberToEdit.name" />
+      <input type="text" class="form-input" placeholder="Enter Email" v-model="memberToEdit.email" />
     </main>
     <button @click="save" class="btn dark bottom-btn create-btn">Save</button>
   </section>
 </template>
 
-<script>
-export default {
-  name: 'member-edit',
-  props: {
-    member: { type: Object },
-  },
-  data() {
-    return {
-      memberToEdit: null,
-      isEdit: false,
-    };
-  },
-  methods: {
-    save() {
-      this.$emit('save', this.memberToEdit);
-    },
-  },
-  created() {
-    this.memberToEdit = this.member;
-    this.isEdit = this.memberToEdit.name ? true : false;
-  },
-};
+<script setup>
+import { ref } from 'vue';
+const props = defineProps({
+  member: { type: Object },
+});
+const emit = defineEmits(['save']);
+
+const memberToEdit = ref(props.member);
+const isEdit = ref(!!memberToEdit.value?.name);
+
+// FUNCTIONS
+function save() {
+  emit('save', memberToEdit.value);
+}
 </script>
