@@ -1,6 +1,6 @@
 <template>
   <section class="main-app-container">
-    <app-header :class="{load: isLoading}" />
+    <app-header />
     <router-view class="main-container" />
     <div v-if="isLoading" class="loader-container screen">
       <div class="app-loader"></div>
@@ -8,21 +8,16 @@
   </section>
 </template>
 
-<script>
-import { appHeader } from '@/modules/common/cmps';
-export default {
-  name: 'App',
-  computed: {
-    isLoading() {
-      return this.$store.getters.isLoading;
-    },
-  },
-  created() {
-    this.$store.dispatch({ type: 'commonStore/loadCurrencies' });
-  },
-  components: {
-    appHeader,
-  },
-};
-</script>
+<script setup>
+import { computed } from 'vue';
+import { useCommonStore } from './modules/common/store';
+import appHeader from '@/modules/common/cmps/app-header';
 
+const commonStore = useCommonStore();
+const isLoading = computed(() => commonStore.isLoading);
+
+// CREATED
+(async function created() {
+  await commonStore.loadCurrencies();
+})();
+</script>
