@@ -1,28 +1,22 @@
+import { defineStore } from 'pinia';
 import { currencyService } from '../services/currency.service.js'
 import { loggerService } from '@/modules/common/services/logger.service.js'
 
-export default {
-  namespaced: true,
-  state: {
+export const useCommonStore = defineStore('common', {
+  state: () => ({
+    isLoading: false,
     currencies: []
-  },
+  }),
   getters: {
-    currencies({ currencies }) { return currencies },
-    currencyCodes({ currencies }) { return Object.keys(currencies)}
-  },
-  mutations: {
-    setCurrencies(state, { currencies }) {
-      state.currencies = currencies
-    }
+    currencyCodes({ currencies }) { return Object.keys(currencies) },
   },
   actions: {
-    async loadCurrencies({ commit }) {
+    async loadCurrencies() {
       try {
-        const currencies = await currencyService.getCurrencyData();
-        commit({ type: 'setCurrencies', currencies })
+        this.currencies = await currencyService.getCurrencyData();
       } catch (err) {
         loggerService.error(err)
       }
     }
   },
-};
+});
