@@ -1,87 +1,73 @@
+import { defineStore } from 'pinia';
+import { useCommonStore } from '@/modules/common/store';
 import { groupService } from '../services/group.service.js'
 import { loggerService } from '@/modules/common/services/logger.service.js'
 
 
-const initialState = () => {
-  return {
-    groups: []
-  }
-}
 
-export default {
-  namespaced: true,
-  state: initialState(),
-  getters: {
-    groups({ groups }) { return groups }
-  },
-  mutations: {
-    setGroups(state, { groups }) {
-      state.groups = groups
-    },
-    addGroup(state, { group }) {
-      state.groups.push(group)
-    },
-    updateGroup(state, { group }) {
-      const idx = state.groups.findIndex(g => g.id === group.id)
-      if (idx === -1) return
-      state.groups.splice(idx, 1, group)
-    },
-    removeGroup(state, { groupId }) {
-      state.groups = state.groups.filter(group => group.id !== groupId)
-    },
-    resetState(state) {
-      state = initState()
-    }
-  },
+export const useGroupStore = defineStore('group', {
+  state: () => ({
+    groups: []
+  }),
   actions: {
-    async loadGroups({ commit }, { filterBy }) {
-      commit({ type: 'setLoading', isLoading: true }, { root: true })
+    async loadGroups({ filterBy }) {
+      const commonStore = useCommonStore()
+      commonStore.isLoading = true
       try {
         // set filterBy 
-        const groups = await groupService.getGroups(filterBy)
-        commit({ type: 'setGroups', groups })
+        this.groups = await groupService.getGroups(filterBy)
+        return this.groups
       } catch (err) {
         loggerService.error(err)
       } finally {
-        commit({ type: 'setLoading', isLoading: false }, { root: true })
+        commonStore.isLoading = false
       }
     },
-    async getGroupById({ commit }, { groupId }) {
-      commit({ type: 'setLoading', isLoading: true }, { root: true })
+    async getGroupById({ groupId }) {
+      const commonStore = useCommonStore()
+      commonStore.isLoading = true
       try {
         return await groupService.getGroupById(groupId)
       } catch (err) {
         loggerService.error(err)
       } finally {
-        commit({ type: 'setLoading', isLoading: false }, { root: true })
+        commonStore.isLoading = false
       }
     },
-    async saveGroup({ commit }, { group }) {
-      commit({ type: 'setLoading', isLoading: true }, { root: true })
+    async saveGroup({ group }) {
+      const commonStore = useCommonStore()
+      commonStore.isLoading = true
       try {
-        const type = (!!group.id) ? 'updateGroup' : 'addGroup'
-
         const savedGroup = await groupService.saveGroup(group)
-        commit({ type, group: savedGroup })
+        
+        if (!!group.id) {
+          const idx = this.groups.findIndex(g => g.id === group.id)
+          if (idx === -1) return
+          this.groups.splice(idx, 1, group)
+        } else {
+          this.groups.push(group)
+        }
+        
         return savedGroup
       } catch (err) {
         loggerService.error(err)
       } finally {
-        commit({ type: 'setLoading', isLoading: false }, { root: true })
+        commonStore.isLoading = false
       }
     },
-    async removeGroup({ commit }, { groupId }) {
-      commit({ type: 'setLoading', isLoading: true }, { root: true })
+    async removeGroup({ groupId }) {
+      const commonStore = useCommonStore()
+      commonStore.isLoading = true
       try {
         await groupService.removeGroup(groupId)
-        commit({ type: 'removeGroup', groupId })
-
+        this.groups = this.groups.filter(group => group.id !== groupId)
+        
       } catch (error) {
         loggerService.error(err)
       } finally {
-        commit({ type: 'setLoading', isLoading: false }, { root: true })
-
+        commonStore.isLoading = false
+        
       }
     }
-  },
-};
+  }
+})
