@@ -1,5 +1,4 @@
-import Vue from 'vue';
-import VueRouter from 'vue-router';
+import { createRouter, createWebHashHistory } from 'vue-router'
 
 import { authRoutes } from '@/modules/auth/routes'
 import { commonRoutes } from '@/modules/common/routes'
@@ -7,7 +6,6 @@ import { groupRoutes } from '@/modules/group/routes'
 import { expenseRoutes } from '@/modules/expense/routes'
 
 
-Vue.use(VueRouter)
 
 const routes = [
   {
@@ -21,26 +19,25 @@ const routes = [
 ]
 
 
-export const router = new VueRouter({
-  mode: 'hash',
-  base: process.env.BASE_URL,
+export const router = createRouter({
+  history: createWebHashHistory(process.env.BASE_URL),
   routes,
 });
 
 
-let _store;
-export const initStore = store => _store = store;
+// let _store;
+// export const initStore = store => _store = store;
 
 
-router.beforeEach(async (to, from, next) => {
-  let loggedInUser = _store.getters['authStore/loggedInUser'];
-  if (!loggedInUser) {
-    if (['login-signup'].includes(to.name)) return next();
-    return router.push('/auth').catch(() => { });
-  } else {
-    if (['login-signup'].includes(to.name)) {
-      return next('/')
-    }
-  }
-  next()
-})
+// router.beforeEach(async (to, from, next) => {
+//   let loggedInUser = _store.getters['authStore/loggedInUser'];
+//   if (!loggedInUser) {
+//     if (['login-signup'].includes(to.name)) return next();
+//     return router.push('/auth').catch(() => { });
+//   } else {
+//     if (['login-signup'].includes(to.name)) {
+//       return next('/')
+//     }
+//   }
+//   next()
+// })
